Refactor AggregationHandle to break cyclic dependency
